restructure C++ code
