+++ csrc/include/CircleDetection/operations.h
